Fixes #7
The documentation indicates that the default behavior of an ACK is the
swapping of the src/dst fields and the echoing of command1 and command2.
Originally, the ACK logic matched on command1 and discarded if the ACK
message did not match command1.  However, further documentation
indicates that both command1 and command2 can be used to convey response
data in an ACK.  Therefore, the only way to match an ACK is by the ACK
flag and the source/destination fields.